move time distance diagram plotting to separate function
