+++ src/ess/wfm/frames_analytical.py
@@ -4,28 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.patches import Rectangle
 import scipp as sc
-
-
-def _angular_frame_edge_to_time(angular_frequency, angle, phase):
-    """
-    Convert an angle on a rotating chopper to a time point (in microseconds).
-    """
-    div = angular_frequency * (1.0 * sc.units.s)
-    return (angle + phase) / div * (1.0e6 * sc.units.us)
-
-
-def _get_frame(frame):
-    """
-    Get coordinates of a chopper frame opening in time and distance.
-    """
-    dist = sc.norm(frame["position"])
-    tstart = _angular_frame_edge_to_time(frame["angular_frequency"],
-                                         frame["opening_angles_open"],
-                                         frame["phase"])
-    tend = _angular_frame_edge_to_time(frame["angular_frequency"],
-                                       frame["opening_angles_close"],
-                                       frame["phase"])
-    return dist, tstart, tend
+from .tools import get_frame_properties
 
 
 def frames_analytical(data, wfm_chopper_names=["WFMC1", "WFMC2"]):
@@ -95,7 +74,7 @@ def frames_analytical(data, wfm_chopper_names=["WFMC1", "WFMC2"]):
 
         # Get frame parameters
         frame = data.meta["choppers"].value["frame", i]
-        dist, tstart, tend = _get_frame(frame)
+        dist, tstart, tend = get_frame_properties(frame)
 
         # Find deltat for the min and max wavelengths:
         # - dt_max is equal to the time width of the WFM choppers windows
@@ -160,159 +139,3 @@ def frames_analytical(data, wfm_chopper_names=["WFMC1", "WFMC2"]):
     #         fig.savefig(plot, bbox_inches='tight')
 
     return frames
-
-
-def _plot(data, frames):
-    """
-    Plot the time-distance diagram that was used to compute the frame
-    boundaries.
-    """
-
-    # x0, x1, y0 = _get_source_pulse(instrument["pulse_length"],
-    #                                instrument["pulse_t_0"])
-
-    # Find detector pixel furthest away from source
-    pos_norm = sc.norm(data.meta["position"])
-    source_pos = sc.norm(data.meta["source_position"])
-    det_last = sc.max(pos_norm)
-    fig, ax = plt.subplots(1, 1, figsize=(9, 7))
-    ax.grid(True, color='lightgray', linestyle="dotted")
-    ax.set_axisbelow(True)
-    psize = det_last.value / 50.0
-    # rect = Rectangle((x0.value, y0.value),
-    #                  x1.value - x0.value,
-    #                  -psize,
-    #                  lw=1,
-    #                  fc='orange',
-    #                  ec='k',
-    #                  hatch="////",
-    #                  zorder=10)
-    ax.add_patch(
-        Rectangle(
-            (0, source_pos.value),
-            (2.0 * data.meta["pulse_t_0"] + data.meta["pulse_length"]).value,
-            -psize,
-            lw=1,
-            fc='lightgrey',
-            ec='k',
-            zorder=10))
-    ax.add_patch(
-        Rectangle((data.meta["pulse_t_0"].value, source_pos.value),
-                  data.meta["pulse_length"].value,
-                  -psize,
-                  lw=1,
-                  fc='grey',
-                  ec='k',
-                  zorder=11))
-    ax.text(data.meta["pulse_t_0"].value,
-            -psize,
-            "Source pulse ({} {})".format(data.meta["pulse_length"].value,
-                                          data.meta["pulse_length"].unit),
-            ha="left",
-            va="top",
-            fontsize=6)
-
-    for i in range(
-            data.meta["choppers"].value["opening_angles_open"].sizes["frame"]):
-
-        dist, xstart, xend = _get_frame(data.meta["choppers"].value["frame",
-                                                                    i])
-        # dist, xstart, xend = _get_frame(i, instrument)
-
-        for j in range(data.meta["choppers"].value["opening_angles_open"].
-                       sizes["chopper"]):
-            ax.plot([xstart["chopper", j].value, xend["chopper", j].value],
-                    [dist["chopper", j].value] * 2,
-                    color="C{}".format(i))
-            if i == data.meta["choppers"].value["opening_angles_open"].sizes[
-                    "frame"] - 1:
-                ax.text((2.0 * xend["chopper", j].data -
-                         xstart["chopper", j]).value,
-                        dist["chopper", j].value,
-                        data.meta["choppers"].value["names"]["chopper",
-                                                             j].value,
-                        ha="left",
-                        va="center")
-
-        col = "C{}".format(i)
-        # left_edge = frames["left_edges"]["frame", i]
-        # right_edge = frames["right_edges"]["frame", i]
-        frame = frames["frame", i]
-        # right_edge = frames["right_edges"]["frame", i]
-        pos = pos_norm.copy()
-        for dim in data.meta["position"].dims:
-            # left_edge = left_edge[dim, 0]
-            # right_edge = right_edge[dim, 0]
-            frame = frame[dim, 0]
-            pos = pos[dim, 0]
-        ax.fill([(data.meta["pulse_t_0"] + frame["right_dt"]).value,
-                 (data.meta["pulse_t_0"] + data.meta["pulse_length"] -
-                  frame["left_dt"]).value,
-                 (frame["left_edges"] + (0.5 * frame["left_dt"].data)).value,
-                 (frame["right_edges"] -
-                  (0.5 * frame["right_dt"].data)).value],
-                [source_pos.value, source_pos.value, pos.value, pos.value],
-                alpha=0.3,
-                color=col)
-        ax.fill([
-            data.meta["pulse_t_0"].value,
-            (data.meta["pulse_t_0"] + frame["right_dt"]).value,
-            (frame["right_edges"] + (0.5 * frame["right_dt"].data)).value,
-            (frame["right_edges"] - (0.5 * frame["right_dt"].data)).value
-        ], [source_pos.value, source_pos.value, pos.value, pos.value],
-                alpha=0.15,
-                color=col)
-        ax.fill([(data.meta["pulse_t_0"] + data.meta["pulse_length"] -
-                  frame["left_dt"]).value,
-                 (data.meta["pulse_t_0"] + data.meta["pulse_length"]).value,
-                 (frame["left_edges"] + (0.5 * frame["left_dt"].data)).value,
-                 (frame["left_edges"] - (0.5 * frame["left_dt"].data)).value],
-                [source_pos.value, source_pos.value, pos.value, pos.value],
-                alpha=0.15,
-                color=col)
-
-        # Minimum wavelength
-        ax.plot([(data.meta["pulse_t_0"] + data.meta["pulse_length"]).value,
-                 (frame["left_edges"] + (0.5 * frame["left_dt"].data)).value],
-                [source_pos.value, pos.value],
-                color=col,
-                lw=1)
-        ax.plot([(data.meta["pulse_t_0"] + data.meta["pulse_length"] -
-                  frame['left_dt']).value,
-                 (frame["left_edges"] - (0.5 * frame["left_dt"].data)).value],
-                [source_pos.value, pos.value],
-                color=col,
-                lw=1)
-        # Maximum wavelength
-        ax.plot([
-            data.meta["pulse_t_0"].value,
-            (frame["right_edges"] - (0.5 * frame["right_dt"].data)).value
-        ], [source_pos.value, pos.value],
-                color=col,
-                lw=1)
-        ax.plot([(data.meta["pulse_t_0"] + frame["right_dt"]).value,
-                 (frame["right_edges"] +
-                  (0.5 * frame["right_dt"].data)).value],
-                [source_pos.value, pos.value],
-                color=col,
-                lw=1)
-
-        # ax.plot([(instrument["pulse_t_0"] + instrument["pulse_length"]).value,
-        #          left_edge.value], [source_pos.value, pos.value],
-        #         color=col,
-        #         lw=1)
-        # ax.text(0.5 * (left_edge + right_edge).value,
-        #         pos.value,
-        #         "Frame {}".format(i + 1),
-        #         ha="center",
-        #         va="top")
-
-    ax.plot([0, sc.max(frames["right_edges"].data).value],
-            [det_last.value] * 2,
-            lw=3,
-            color='grey')
-    ax.text(0.0, det_last.value, "Detector", va="bottom", ha="left")
-    ax.set_xlabel("Time [microseconds]")
-    ax.set_ylabel("Distance [m]")
-
-    return fig
